0.9.0 - brought in line with new component-boilerplate
