+++ response-items/freetext-component/package.json
@@ -1,14 +1,12 @@
 {
   "name": "freetext",
-  "bundle": "FreeText",
-  "version": "0.8.0",
+  "componentName": "FreeText",
+  "version": "0.9.0",
   "description": "The FreeText Question Component that ships with the DECSYS Survey Platform",
-  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "rollup": "rollup -c",
-    "esm": "node ./build/esm-export",
-    "build": "npm run rollup && npm run esm",
+    "build": "npm run rollup",
     "storybook": "start-storybook -p 6006",
     "build-storybook": "build-storybook"
   },
@@ -16,26 +14,27 @@
   "license": "UNLICENSED",
   "devDependencies": {
     "@babel/core": "^7.4.0",
-    "@storybook/addon-actions": "^5.0.3",
-    "@storybook/addon-knobs": "^5.0.3",
+    "@storybook/addon-actions": "^5.0.11",
+    "@storybook/addon-knobs": "^5.0.11",
     "@storybook/addon-links": "^5.0.3",
     "@storybook/addons": "^5.0.3",
-    "@storybook/react": "^5.0.3",
-    "babel-loader": "^8.0.5",
-    "babel-preset-env": "^1.7.0",
-    "babel-preset-react": "^6.24.1",
+    "@storybook/react": "^5.0.11",
+    "@babel/preset-env": "^7.4.4",
+    "@babel/preset-react": "^7.0.0",
     "prop-types": "^15.7.2",
     "react": "^16.8.4",
     "rollup": "^1.7.0",
     "rollup-plugin-babel": "^4.3.2",
     "rollup-plugin-commonjs": "^9.2.1",
+    "rollup-plugin-json": "^4.0.0",
     "rollup-plugin-jsx": "^1.0.3",
     "rollup-plugin-node-resolve": "^4.0.1",
     "rollup-plugin-replace": "^2.1.1",
+    "rollup-plugin-terser": "^4.0.4",
     "styled-components": "^4.1.3"
   },
   "dependencies": {
-    "@decsys/param-types": "^0.2.0",
+    "@decsys/param-types": "^0.3.1",
     "@smooth-ui/core-sc": "^10.0.2",
     "styled-icons": "^7.6.0"
   },
