+++ response-items/freetext-component/rollup.config.js
@@ -2,16 +2,27 @@ import resolve from "rollup-plugin-node-resolve";
 import cjs from "rollup-plugin-commonjs";
 import replace from "rollup-plugin-replace";
 import babel from "rollup-plugin-babel";
+import json from "rollup-plugin-json";
+import { terser } from "rollup-plugin-terser";
 
 const pkg = require("./package.json");
+
+// the actual module exports from the bundled IIFE
+const footer = `
+export const name = DecsysComponent.displayName;
+export default DecsysComponent;
+`;
 
 export default {
   input: "src/index.js",
   output: {
     format: "iife",
     name: "DecsysComponent",
-    file: `dist/${pkg.bundle}.js`,
-    exports: "named", // explicitly set default export as a `<name>.default` named export
+    file: `dist/${pkg.componentName}.js`,
+    sourcemap: true,
+    preferConst: true,
+    compact: true,
+    footer: footer,
     globals: {
       react: "React",
       "styled-components": "styled",
@@ -37,7 +48,9 @@ export default {
         "@babel/preset-react"
       ]
     }),
-    resolve(),
-    cjs()
+    resolve({ preferBuiltins: false }),
+    cjs(),
+    json(),
+    terser()
   ]
 };
