+++ response-items/freetext-component/src/metadata.js
@@ -1,0 +1,3 @@
+import { AlignLeft } from "styled-icons/fa-solid/AlignLeft";
+
+export const icon = <AlignLeft />;
